chat_bot/requirements.txt
google-genai
python-dotenv
lmnr
requests
